+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ##CHANGELOG
 
+1.1.0 -
+- added a settings window (accessible from the stock application launcher or blizzy78's toolbar).
+- added a "only deploy when safe" setting to prevent staged chutes from deploying until it's safe to do so.
+
 1.0.2 - 2015/08/06
 - added a new "clearBackground" setting to the config file which will completely remove the grey gradient texture from the stack icon background.
 Setting this to "true" should make it easier for color blind users to find a workable color combination.